// src/components/MainListInfoForm/type.ts
export enum FIELD_NAMES {
    STATUS = 'status',
    LIST_NAME = 'listName',
    DESCRIPTION = 'description',
    VISIBILITY = 'visibility',
    RECORD_TYPE = 'recordType'
}

export enum VISIBILITY_VALUES {
    PRIVATE = 'private',
    SHARED = 'shared'
}

export enum STATUS_VALUES {
    ACTIVE = 'active',
    INACTIVE = 'inactive'
}